requirements.txt
snet-cli
pyopenssl
